+++ packages/global-styles-ui/src/screen-typography.tsx
@@ -3,7 +3,6 @@ import { __experimentalVStack as VStack } from '@wordpress/components';
 import { useContext } from '@wordpress/element';
 import { ScreenHeader } from './screen-header';
 import { ScreenBody } from './screen-body';
-import TypographyElements from './typography-elements';
 import TypographyVariations from './variations/variations-typography';
 import FontFamilies from './font-families';
 import FontSizesCount from './font-sizes/font-sizes-count';
@@ -17,14 +16,13 @@ function ScreenTypography() {
 			<ScreenHeader
 				title={ __( 'Typography' ) }
 				description={ __(
-					'Available fonts, typographic styles, and the application of those styles.'
+					'Manage the fonts and typographic styles available across the site.'
 				) }
 			/>
 			<ScreenBody>
 				<VStack spacing={ 7 }>
 					<TypographyVariations title={ __( 'Typesets' ) } />
 					{ fontLibraryEnabled && <FontFamilies /> }
-					<TypographyElements />
 					<FontSizesCount />
 				</VStack>
 			</ScreenBody>